Example modelblueprint objects

A family of constructor functions that return ready-to-use modelblueprint objects for common R model types. All examples use the built-in mtcars dataset with a reproducible 75/25 train/test split.

Usage

mb_lm_regression()

mb_lm_classification()

mb_glm_regression()

mb_glm_binomial()

mb_glm_poisson()

mb_rpart_regression()

mb_rpart_classification()

mb_rf_regression()

mb_rf_classification()

mb_xgb_regression()

mb_xgb_classification()

mb_h2o_regression()

mb_h2o_classification()

mb_h2o_glm_large(n = 5000L)

Arguments

n

[integer(1)] Number of rows to simulate. Default 5000L is enough to see the caching benefit of the dashboard. Use 50000L or more to stress-test prediction latency on large models.

Value

A modelblueprint object.

Details

Regression target: mpg (continuous)

Classification target: vs (binary 0/1)

Functions that wrap optional packages (rpart, randomForest, xgboost, h2o) will error informatively if the package is not installed. H2O functions also require a Java installation and will start an H2O cluster automatically.

post_process_fun is set on classification models that return a probability matrix (rpart, randomForest) so that predict() always returns a plain numeric vector of P(class = 1).

Examples

mb <- mb_lm_regression()
predict(mb, mtcars)
#>  [1] 22.32234 21.63226 24.79266 20.71217 17.42918 20.15495 15.59625 23.09422
#>  [9] 22.50423 19.82821 19.82821 15.61850 16.53859 16.40329 11.89625 11.21379
#> [17] 11.11019 25.68869 27.56802 26.69760 24.31563 17.74166 17.97168 14.86559
#> [25] 16.33318 26.40583 25.32209 26.55336 16.27670 20.54108 13.69195 23.20928
one_way(mb, var = "wt")

mb_cls <- mb_glm_binomial()
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
predict(mb_cls, mtcars)   # returns probabilities in 0 to 1
#>  [1] 2.122640e-02 6.087156e-01 7.219865e-01 9.999917e-01 2.666064e-10
#>  [6] 1.000000e+00 2.220446e-16 1.000000e+00 1.000000e+00 9.996542e-01
#> [11] 9.996542e-01 5.717862e-07 1.918360e-09 4.434027e-09 1.195615e-04
#> [16] 6.873054e-06 3.177700e-10 9.999995e-01 9.999973e-01 9.998752e-01
#> [21] 9.986928e-01 1.879847e-03 4.530718e-04 2.220446e-16 2.361581e-07
#> [26] 9.999584e-01 2.875166e-01 3.373855e-11 2.220446e-16 2.220446e-16
#> [31] 2.220446e-16 3.605970e-01